requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.7
django-allauth==0.58.2
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
